Demo/requirements.txt
flask
flask-cors
torch
transformers
accelerate
pymongo
python-dotenv
sentencepiece
huggingface-hub